aim22-reverseisp/teams/HIT-IIL/sRGB-to-RAW-s7/requirements.txt
absl-py==1.2.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
colour-demosaicing==0.1.6
colour-science==0.3.16
cupy-cuda102==10.6.0
cycler==0.11.0
einops==0.4.1
fastrlock==0.8
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.5
importlib-metadata==4.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.5.2
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
rawpy==0.17.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tifffile==2021.11.2
timm==0.6.5
torch==1.8.1
torchvision==0.9.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
